feat(add-mix-cluster-feature): refactor cityMaps
