+++ src/components/uiCityMaps/CityMapSydney.js
@@ -2,8 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { getRandomPoint } from './_helpers'
 import { SingleOccupancy, MultiOccupancy } from 'pcln-icons'
-import { MapItemContainer } from '../../GoogleSuperCluster'
-import GoogleSuperCluster from '../../GoogleSuperCluster'
+import GoogleSuperCluster, { MapItemContainer } from '../GoogleSuperCluster'
 
 const CityMapSydney = ({ hotels, defaultZoom, isZoomOut }) => {
   const colors = []
